refactor(core): 提取 memory recall 工具到 memory/recall.rs
- 将 recall_memory 注入、执行、重复拦截等函数提取到 memory/recall.rs
- attachment 分析工具保留在 core/mod.rs（依赖 RuntimeEngine 多模态客户端）
- 修复 recall.rs 中误混入的 attachment 代码
- core/mod.rs 进一步降至约 1260 行
