+++ crates/tiangong-core/src/memory/mod.rs
@@ -1,3 +1,4 @@
 pub mod gui_api;
+pub mod recall;
 pub mod registry;
 pub mod turn_result;
